+++ src/flask_exts/template/dataset.py
@@ -1,9 +1,0 @@
-class Dataset:
-    def __init__(self):
-        self.admin_base_template = "admin/base.html"
-        self.title = {
-            "view": "View",
-            "edit": "Edit",
-            "delete": "Remove",
-            "new": "Create",
-        }