+++ tests/template/bootstrap5/conftest.py
@@ -1,12 +1,12 @@
 import pytest
 from flask import Flask
 from flask_exts import Manager
-from flask_exts.template.theme import BootstrapTheme
+from flask_exts.template.theme import Theme
 from flask_exts.template.base import Template
 
 class MyTemplate(Template):
      def get_theme(self):
-          return BootstrapTheme(5)
+          return Theme()
 
 class MyManager(Manager):
     def get_template(self):